Add a new feature to track missing fallouts
As reported, the mailing list pkg-fallout is not receiving/storing all
the pkg-fallouts.  This adds a management command that, by default,
searches the builds from the last thirty days, and goes to the Poudriere
build log of errors to check if all errors there are stored in the
portsfallout database. If they are not, it processes the log file and
inserts it into the database, without the report URL (which is the URL
from the mailing archive).

Fixes #5